test/http/middleware/RequestDataMiddleware.test.ts
import 'jest'
import * as Sinon from 'sinon'
import { RequestDataMiddleware } from '../../../lib/http/middleware/RequestDataMiddleware'
import { isPromise } from '../../../lib/private/isPromise'
import { ViewResponse } from '../../../lib/http/response/types/ViewResponse'
import { HandlebarsHelper } from '../../../lib/view/handlebars/HandlebarsHelper'
import { HandlebarsViewResponse } from '../../../lib/view/handlebars/HandlebarsViewResponse'

describe('RequestDataMiddleware', function() {
  it('is fit for najs-binding', function() {
    expect(RequestDataMiddleware.className).toEqual('Najs.RequestDataMiddleware')
  })

  describe('.after()', function() {
    it('returns a promise', async function() {
      const request = {}
      const response = {}
      const controller = {}
      const result = {}
      const instance = new RequestDataMiddleware()

      const returnValue = instance.after(<any>request, <any>response, result, <any>controller)
      expect(isPromise(returnValue)).toBe(true)
    })

    it('does nothing if the view is ViewResponse', async function() {
      const request = {}
      const response = {}
      const controller = {}
      const result = new ViewResponse('test')
      const instance = new RequestDataMiddleware()

      const returnValue = await instance.after(<any>request, <any>response, result, <any>controller)
      expect(returnValue === result).toBe(true)
    })

    it('calls result.helper and add "Input", "Body", "Query", "Params" helpers if the view is HandlebarsViewResponse', async function() {
      const request = {}
      const response = {}
      const controller = {}
      const result = new HandlebarsViewResponse('test')

      const helper = () => {}
      const createHelperStub = Sinon.stub(HandlebarsHelper, 'create')
      createHelperStub.returns(helper)

      const instance = new RequestDataMiddleware()

      expect(result.getVariables()).toEqual({})
      const returnValue = await instance.after(<any>request, <any>response, result, <any>controller)
      expect(result.getVariables()).toEqual({ helpers: { Input: helper, Body: helper, Query: helper, Params: helper } })
      expect(returnValue === result).toBe(true)

      expect(createHelperStub.callCount).toEqual(4)

      createHelperStub.restore()
    })
  })
})
